lan-file-share/requirements.txt
fastapi
uvicorn
python-multipart
qrcode[pil]
python-dotenv
jinja2
